Initial Commit
I put both the class definitions in the same file initially. When I
make the eventual package itself, I will most likely differentiate
between the Trap class and the Simp class. I believe that this is a
correct way to set myself up for actually calculating Simpson’s Rule
and the Trapezoidal Rule by making the class definitions themselves
require the coordinates of the starting and ending points of a curve,
as I specified in my comments.